Immediately give the user some feedback about the password reset, then confirm it
I know it's kind of weird to set the background color in the store, but
since the alert type is a result of what action's been emitted, it
seemed not super inappropriate. I'm kind of going outon a limb here and
I wanna know what everyone thinks. I like eliminating an if-block in the
component, but it feels weird to have a className in the store.
¯\_(ツ)_/¯

Use null instead of false to invalidate confirmation text