homepage design cleanup 1
Updated header to blue, which also means I had to update the gif to
include a blue background to remove the white pixels around it. Copy
was updated to reflect what was currently on the site (tho maybe I
didn’t need to do this). Added smaller fontsize for mobile breakpoint.
Added hover for footer links. Adjusted font sizes and padding. Moved
the body of text to the left and images to the right, as that felt
easier to digest. There are some things that Pletcher updated that I
didn’t update here. Just FYI. Also, I wasn’t sure how to create a new
.sm padding. I’m ready for your comments on my hack :)